Introduce surfaceflinger_end2end_tests
Introduces a minimal surfaceflinger_end2end_test target, which will work
by controlling both the "front-end" of surfaceFlinger via the libgui
ISurfaceComposer.aidl interface, as well the "back-end" of
surfaceflinger via the hardware Composer3 IComposer.aidl interface.

There is only a minimal implementation for everything. To start out, the
test registers a simple "fake" composer3 backend service that asserts if
SurfaceFlinger actually uses it to create a client interface, and
otherwise the initial test just tries to connect to the front end
service.

The main purpose of this CL is to introduce all the main framework
classes, and otherwise set up the build target.

For now this fails, with SurfaceFlinger establishing a connection to the
service, but the test process aborts in the stub to create the client
interface. SurfaceFlinger doesn't crash, but is in a bad state. The
AndroidTest.xml file however brings the target back to a normal working
state with tear-down commands meant for this purpose.

Flag: TEST_ONLY
Bug: 372735083
Test: atest surfaceflinger_end2end_tests
Change-Id: I78be88a3eb3df9b6c700f78bdef66f9dc71399c7