Add test to verify the repeated key when window focus changed
In previous version, when an user long press a key and it's not consumed
by the policy, a synthesized repeat key will be delivered to the focused
window even it changed to another one.

This will add a new test to verify it won't synthesize the new repeat
key when the focused window is changed.

Flag: EXEMPT bugfix
Bug: 400516905
Test: atest inputflinger_tests
Change-Id: Iafc85800bfc292bc1daab6d70c67b8597757bac1