SF: Cache display port in DisplayIdentification on onHotplugConnect
In order to be able to remove direct dependencies on a DisplayId's int
value, we have to find alternatives for the API it provides. One such
piece of data is a display port, which is encoded directly in the ID int
value.

Cache the port as it is retrieved from HWC into DisplayIdentification so
it can be served later via alternative ISurfaceComposer APIs.

Bug: 374163881
Bug: 377307639
Flag: EXEMPT refactor
Test: libsurfaceflinger_unittest
Change-Id: If255418ed4047531417004c4a1fd710061d93685