SF: Allow debug.sf.hwc_service_name to be fully-qualified
The checks in ComposerHal and AidlComposerHal for whether or not to use
the AIDL service interface or fall back to the HIDL service interface
just checked if the name could qualified using the AIDL service name
prefix, and matched to a name in the VINTF manifest. If it was not
found, the name was assumed to be a HIDL interface.

This check meant that a test that registers its own AIDL service and
sets debug.sf.hwc_service_name appropriately cannot get SurfaceFlinger
to load it.

This change allows the string to be set to a fully qualified name, like
"android.hardware.graphics.composer3.IComposer/fake", and
AidlComposerHal will now recognize from the prefix that it is meant to
be an AIDL implementation, and so it should be used. Importantly this
also skips the check that the name is registered, which is problematic
for a test.

The old functionality is kept. Setting the string to "default" will
cause a check to see if
"android.hardware.graphics.composer3.IComposer/default" is registered,
and use it if found, or otherwise fall back to the HIDL "default"
interface if not.

Flag: EXEMPT for use by tests only
Bug: 372735083
Test: m and run cf target

Change-Id: I26a0d92d1600a48facb75f956400433d03efcf5e