SF: Cache display port in physical DisplaySnapshot
In order to be able to remove direct dependencies on a DisplayId's int
value, we have to find alternatives for the API it provides. One such
piece of data is a display port, which is encoded directly in the ID int
value.

Cache the port as it is retrieved from HWC into DisplaySnapshot when
they are updated or created so it can be served later via alternative
ISurfaceComposer APIs.

Bug: 374163881
Bug: 377307639
Flag: EXEMPT refactor
Test: libsurfaceflinger_unittest
Change-Id: I7a7e38695ca6065238644457986f12dcce4b59d0