Add pre-built for satellite feature XML
Bug: 351819362

The satellite feature XML is shared by all OEMs producing a device with the satellite feature being returned as a system feature from the package manager. Currently it can only be included by the deprecated practice of adding to PRODUCT_COPY_FILES, however. This change adds a prebuilt_etc declaration for it so it can be included by adding to PRODUCT_PACKAGES instead.

Test: add android.hardware.telephony.satellite.prebuilt.xml to PRODUCT_PACKAGES and build
Flag: EXEMPT refactor
Change-Id: I39d001dba6c5e0121fa269339e00102a2e247539