SF: Add EDID-ID Fabrication Logic
Adds logic to fabricate a display ID that is based solely on a given,
pre-parsed EDID data.

It does so by concatenating the EDID's generic fields, such as
manufacturer ID, product ID, etc. as a string, and hash-combine them
with the hashed values of the display serial numbers.

All hash functions are stable in order to consistently reproduce display
IDs when the same information is fed via the EDID blob.

Flag: com.android.graphics.surfaceflinger.flags.stable_edid_ids
Bug: 366042891
Test: N/A
Change-Id: I7693d4b0ca9fee5ed190c3530a69300a8a530cd2