Moved the Perfetto flag check into SurfaceFlinger.
Having the flag check in Perfetto creates a dependency on the flags, which are currently not available in vendor. This prevents systems in vendor (such as Lyric) from using libtracing_perfetto.

The flag has been 100% enabled at this point (b/303199244), so Perfetto users don't have to worry anymore about the API not being enabled. That said, specific users can choose to do the flag check in their own scope to retain the ability to turn it off if needed, like SurfaceFlinger will now do with this CL.

Bug: 373924379
Change-Id: I761c6f0137c0301cfb8fbac942291863053ee1de
Flag: EXEMPT - moves an existing flag check to a different location