Implement functions for managing Graphite memory budget.
The main fix in this change is to make sure we are actually
setting the memory budget when using Graphite. Previously
we were using the default Graphite budget which was much
higher than what we would normally want.

Additionally, this implemented the purgeUnlcokedScratchResources()
call. This is called when switching between protected and
non-protected contexts.

Finally this fixed dumpMemoryStatistics to also dump memory
held by the Recorder.

Bug: b/293371537, b/385380555
Test: manual comparison of memory metric
Flag: com.android.graphics.surfaceflinger.flags.graphite_renderengine

Change-Id: I02befd7caa0b9eeb90316a082425a9905d91c57b