Allow touchpad debug logs to be enabled on the fly
This is needed so that we can turn them on during particular CTS tests.

The IS_DEBUGGABLE_BUILD declaration in Logging.cpp is a copy of the one
in services/inputflinger/reader/Macros.cpp, but unfortunately we can't
put debugTouchpadGestures there because Macros.h defines LOG_TAG, so we
can't include it from GesturesLogcatAdapter.cpp.

Test: $ atest android.input.cts.hostside.InputAtomsTest
      with the other CLs in this topic patched. Check that Gestures
      library logs appear with the Gestures logcat tag, and "Gesture
      ready:" and "New hardware state:" logs appear with the InputReader
      tag
Bug: 397288324
Flag: EXEMPT log only update
Change-Id: Id92581014836ba856b29741117e9f6a397d9422c