ExternalStylusIntegrationTest: Reduce likelihood for flake
In the test, an event will be generated after a 72ms timeout, and
we assert that no event is generated before that timeout. Since this
is a racy assertion, just wait for 1ms to ensure that an event isn't
immediately generated.

If this is not enough to bring the flakes rates to an acceptable level,
we will need to refactor the tests to remove the race altogether.

Bug: 386801813
Change-Id: Id223930925ce724e216e61a9e68e1da79bed6f04
Test: Presubmit
Flag: TEST_ONLY