Enable searching for mounted public volumes under any user
Currently, InstalldNativeServices when collecting mounted storage to
invalidate, only looks for mounts under /mnt/pass_through/0. This means
that public volumes mounted under users that aren't user 0 are never
collected. This change ensures that all of these public volumes are
collected by checking for /mnt/pass_through/0/emulated and for
/mnt/pass_through/x/UUID, where x is an integer greater than or equal to
0 and UUID is of the format HHHH-HHHH, H being any uppercase
letter or a digit. In this change, the regex checks for /emulated only
under user 0 because /mnt/pass_through/x/emulated/10 points to the same
thing regardless of the value of x, so only checking under user 0, which
is always mounted, prevents killing the same mount twice.

Test: atest ExternalStorageHostTest
Fixes: 299996244
Flag: EXEMPT refactor

Change-Id: Icb4fca7213f9e656594625a83ba16a7ffc8251fe