EventHub: Search IDC for additional lights
When an input device is connected to an internal bus, for example, an
internal keyboard that uses the i8042 protocol, the keyboard backlight
may not be located in the same path as the input device itself.  To
handle this case, this commit adds the functionality for a Input Device
Configuration (IDC) file to be searched for additional lights. This
path, if provided, is then added to the search path for lights.

The new key is named: `device.additionalSysfsLedsNode`.  The value is
the path to the sysfs node for the additional LED.

Bug: 357090960
Flag: EXEMPT bugfix
Test: Manually on device that has an internal keyboard with backlight
        that uses the i8042 keyboard protocol with updated .idc file for
        internal keyboard, verify that keyboard backlight is able to be
        found and associated with the internal keyboard.

Change-Id: I4fbcd05dc738ff7656111c6d1e3484d385c92616