Remove redundant isSplit assignments
Now that touch is always split (except for mouse events), we can
also simplify the assignment of isSplit variable and make it const.

Eventually, we can consider changing the touchpad code so that we can
completely remove the non-split code handling from dispatcher and
simplify this even further, because touchpad scroll should always go to
the location of the cursor.

Bug: 239934827
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST
Flag: EXEMPT refactor
Change-Id: Ib753f798fd77d5d74434b9cbb02f17cd07f5906b