vkjson: Move redefinition of FALLTHROUGH_INTENDED
The macro FALLTHROUGH_INTENDED is defined in both
vulkan/vkjson/vkjson.h and android-base/macros.h, causing build errors
for any other modules that include both header files.

Move the redefinition to vkjson.cc.

Note that <android-base/macros.h> is not available for the libvkjson_ndk
build, so we cannot use that version of the macro and eliminate the
redefinition entirely.

Bug: 372694741
Test: CQ
Flag: EXEMPT bugfix
Change-Id: I02a0f2a50f3807ca932e84a5a002598b04c5f92a