bufferqueues: IConsumerListener/IGraphicBufferConsumer no longer need binder
This CL removes the use of binder from IConsumerListener and
IGraphicBufferConsumer. These interfaces are both only ever called by
BufferQueue and/or BufferQueueConsumer classes and not exposed outside
the process.  As a result, there's no benefit to using binder interfaces
for them, since the process would have to crash if BufferQueue or
BufferQueueConsumer calls methods that are not supported.

This also fixes up a currently-disabled BufferQueue test.

Bug: 389157713
Change-Id: I119839d2441b3a1455ff840dff4154cf54a425fa
Flag: EXEMPT refactor
Test: builds, presubmit