IGpuService: Add getFeatureOverrides()
Add getFeatureOverrides() to the IGpuService interface. This function
returns the list of feature overrides parsed by GpuService to the
caller. This list of feature overrides will be stored by GraphicsEnv, so
it can later be used to override features when creating the ANGLE EGL
display.

Bug: 372694741
Test: CQ
Flag: com.android.graphics.graphicsenv.flags.feature_overrides
Change-Id: Idb727e23306c3deb2b8f98e6db3d9906bb63598f