Add test for simultaneous mouse and touch with transfer
When transferTouch is used while both touch and mouse are active, the
current behaviour is strange. Upon transfer of touch, the transfered-to window
receives HOVER_EXIT, even though it never received HOVER_ENTER or
HOVER_MOVE. Also, the touch gesture is transfered with ACTION_DOWN, but
subsequent touch events are not delivered to the transfered-to window.

In this CL a test is added to reproduce this.

Bug: 382473355
Test: atest --host inputflinger_tests
Flag: TEST_ONLY
Change-Id: Id71415b750ba1c58b5882b87640f4dfc26b79182