Move uses of SkSurface::makeImageSnapshot to makeTemporaryImage.
In graphite, makeImageSnapshot no longer has a copy on write symantic.
Instead it will always make a copy which is not need in many cases.
This CL changes the the uses of makeImageSnapshot where a copy isn't
needed to call makeTemporaryImage instead. This avoids the extra
copies in Graphite.

Besides possible performance wins, this also reduces a lot of extra
memory that was being allocated for these copies. This is most seen
in situtations using a lot of blurs.

Bug: b/293371537 , b/385380555
Test: manual comparison of memory metric
Flag: com.android.graphics.surfaceflinger.flags.graphite_renderengine

Change-Id: Ia9ed5c778f083b7a506158510438e656a4625a67