Mocked device should be consistently external or internal
Currently, we are allowing device's isExternal API to return different
values dynamically. In practice, this property should not change.

In this CL, make this property determined upon construction, and force
the current users to explicitly specify it.

This will allow us to more simply emulate external touchscreens in
subsequent CLs.

Bug: 378308551
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST
Flag: TEST_ONLY
Change-Id: Iaf917c336fba8c66f3b7f7459bf7fe9aaae02d5e