SF: Merge properties that skip boot animation
The flag debug.sf.nobootanimation causes the bootanim to exit early and
not show an animation. When that flag is set, launching bootanim is
pointless, but it slows down boot on average by about 800 ms on some
devices [1]. Quick analysis showed that loading bootanim requires
reading 80 MB of read-only memory [1], which is probably the reason
behind the slowdown during boot.

Optimize that by not launching bootanim at all when
debug.sf.nobootanimation is set, similar to the behavior when
debug.sf.boot_animation is false.

References:
[1] Brya boottime: bootanim takes ~1.5s, b/373711392

Bug: 324366212
Bug: 373711392
Test: Test all 4 combinations of
      setprop debug.sf.boot_animation {false,<null>}
      setprop debug.sf.nobootanimation {1, <null>}
Flag: EXEMPT sysprop alias

Change-Id: I5599072c14008ae67ab8aec4057bda953ac5b978
Signed-off-by: Paz Zcharya <pazz@google.com>