EventHub: Refactor AssociatedDevice creation logic
Refactor the AssociatedDevice creation logic to make it easier to
implement upcoming changes to AssociatedDevice initialization.

In particular, we want to make it simple to differentiate the case where
an existing AssociatedDevice needs to be updated from the case where a
new AssociatedDevice needs to be created.

Bug: 245989146
Bug: 357090960
Test: Presubmit
Flag: EXEMPT refactor
Change-Id: I8d86f56f3c00f9ae13232a60e338ff7ec246ce3b