Add empty, dump methods to BlockingQueue
BlockingQueue can now be dumped (since the user can't access its internal
state, and it's too clunky to be making a copy just so that we can
destructively print it). Also, "empty" method is useful when converting
code from std containers to the BlockingQueue.

At the same time, mark the "size" method as const.

Bug: 376713684
Test: used in another CL in frameworks/base. build only here
Flag: EXEMPT refactor
Change-Id: I3fda00b793b512bdba4c85dbf561dfc5e9724601