[CD Cursor] Fix cursorState lookup for non-primary displays
In connected display scenario touchstate display for mouse is always
identified by primary display. At present
DispatcherTouchState::forTouchAndCursorStatesOnDisplay is the only API
in TouchStates that does not adhere to this mapping. This can cause the
lookup for cursorState to fail.

IsPointerInWindow is the only public API in dispatcher that uses this
and fails. This Cl updates the lookup and adds tests for API called with
the non-primary display in topology.

Test: inputflinger_tests with the flag enabled/disabled
Bug: 395033854
Flag: com.android.input.flags.connected_displays_cursor
Change-Id: I1515171802478a7132206562e41a7586b8ca8559