Surface: Add 'isBufferOwned' call
Camera's StreamSplitter keeps track of buffers and won't attach them if
they're already owned. This allows us to keep that behavior as we move
over from IGBP to surfaces.

Ideally, I think, IGBPs should fail to attach the same buffer multiple
times. Even if not, this call should probably be to the IGBP instead of
the Surface for more accuracy. But this replicates the old behavior.

Bug: 340933206
Flag: com.android.graphics.libgui.flags.wb_stream_splitter
Test: new test

Change-Id: Idbb200202012c9eae2668616dcff277c925c3907