Fix the issue of screen flashing when adjusting ambient light when playing HDR video.
Bug:360380695
Test:Manual

The comparison operation comment of the LayerSettings.h file describes "Keep in sync with custom comparison function in compositionengine/impl/ClientCompositionRequestCache.cpp". However, for the comparison of LayerSetting, white point value comparison is added in LayerSettings.h, but not in ClientCompositionRequestCache.cpp, which makes the HDR video dimming cache ineffective and causes screen flickering.

Change-Id: I54d230ddf176660c5e83f771a59154debce7f541
Signed-off-by: zhengchuanlong <zhengchuanlong@xiaomi.com>