Fix display brightness for HDR layers during transitions
Transitions with HDR would have a white flash due to incorrect
logic with setting display brightness. The bug was caused by
attempting to evaluate the logic for capturedHdrLayers with
several other conditions, causing the brightness to be set
to the sdrWhitePointNits if any of the conditions were false.

Fixes: b/397159125
Test: play HDR video on YouTube and switch orientation to
observe if a white flash occurs
Flag: EXEMPT, bug fix

Change-Id: Ib15060e5b113ec1345da6b3b51a66132c6ae4e6c