Don't skip blurring recently-decomposed cachedsets.
If we had found that a cachedset had a blur effect applied, then we tell the layer requesting the blur to skip blurring. This is so that we do not blur twice.

But, this does not hold if we needed to update a cachedset, meaning that the blur effect was no longer applied. This was previously causing us to never render a blur.

Bug: 390403736
Change-Id: I4389a8c73b20556abe392fe3b464408c6775fc34
Flag: EXEMPT bug fix
Test: Notification shade + blurs