[CD Cursor] Add DisplayTopologyValidator
Add a DisplayTopologyValidator to verify assumptions made in the
PointerChoreographer and InputDispatcher, some sanity checks on the
topology graph to flag issues realted to topology for debugging.

The validator will be used in the InputManager to validated and discard
invalid topology updates.

Test: atest inputflinger_tests
Bug: 401219231
Flag: com.android.input.flags.enable_display_topology_validation

Change-Id: I624bc32b73144cc047f317ef889f7f07cc1c3dfc