InputDispatcher: Provide a way to transfer the entire gesture
In this CL, we modify the behaviour of 'transferTouchGesture' to also
allow the caller to request that the entire gesture is transferred to
the destination window.

Suppose the original window receives ACTION_DOWN and calls
transferTouchGesture to the destination window. Before this CL, the
subsequent pointers would always be routed to the original window.

In this CL, the caller of 'transferTouchGesture' can choose to transfer
the 'entire gesture', meaning that subsequent pointers from the same
gesture that land in the original window would automatically be sent to
the destination window.

Bug: 397979572
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST
Flag: com.android.input.flags.allow_transfer_of_entire_gesture
Change-Id: I760ddb8aa8f329d0b21788e6d6e83c802015ee1d