libbinder: export fewer symbols for vendor builds
See https://r.android.com/3087749 for details.

This commit brings the vendor build to parity with the other builds. It
may cause some initial issues for vendors, but that should be more than
compensated by the ongoing issues it will fix for them. For example,
when backporting security fixes, we'd often see ABI breakages because
the set of internal symbols might change (before this commit we'd be
exporting all internal symbols). That should never happen after this
change.

Bug: 338458975
Bug: 349657329
Test: m
Change-Id: I5f77a8d1e16d38c5e83f9b8842368a967b4ee229